pyproject.toml
[project]
name = "trustenroll"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aerich==0.8.2",
    "aiofiles==24.1.0",
    "aiogram>=3.19.0",
    "aioredis==2.0.1",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "async-timeout==5.0.1",
    "asyncclick==8.1.8.0",
    "asyncpg==0.30.0",
    "babel==2.17.0",
    "bcrypt==4.3.0",
    "click==8.1.8",
    "colorama==0.4.6",
    "dictdiffer==0.9.0",
    "fastapi==0.115.11",
    "fastapi-admin==1.0.4",
    "fastapi-cache2>=0.2.2",
    "h11==0.14.0",
    "httptools==0.6.4",
    "idna==3.10",
    "iso8601==2.1.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "passlib>=1.7.4",
    "pendulum==3.0.0",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pypika-tortoise==0.5.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "redis>=5.2.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.46.1",
    "time-machine==2.16.0",
    "tomlkit==0.13.2",
    "tortoise-orm==0.24.2",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "uvicorn==0.34.0",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
]

[tool.aerich]
tortoise_orm = "backend.models.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "ruff>=0.11.2",
]
tg-bot = [
    "aiogram>=3.19.0",
    "python-dotenv>=1.0.1",
]
